+++ src/controller_factory.cpp
@@ -28,6 +28,7 @@
 #include "controller/playstation3_usb_controller.hpp"
 #include "controller/saitek_p2500_controller.hpp"
 #include "controller/saitek_p3600_controller.hpp"
+#include "controller/xeox_controller.hpp"
 #include "controller/wiimote_controller.hpp"
 #include "controller/xbox360_controller.hpp"
 #include "controller/xbox360_wireless_controller.hpp"
@@ -81,6 +82,9 @@ ControllerFactory::create(XPadDevice const& dev_type, libusb_device* dev, Option
 
     case GAMEPAD_SAITEK_P3600:
       return ControllerPtr(new SaitekP3600Controller(dev, opts.detach_kernel_driver));
+
+    case GAMEPAD_XEOX:
+      return ControllerPtr(new XeoxController(dev, opts.detach_kernel_driver));
 
     case GAMEPAD_LOGITECH_F310:
       return ControllerPtr(new LogitechF310Controller(dev, opts.detach_kernel_driver));
@@ -168,6 +172,10 @@ ControllerFactory::create_multiple(XPadDevice const& dev_type, libusb_device* de
 
     case GAMEPAD_SAITEK_P3600:
       lst.push_back(ControllerPtr(new SaitekP3600Controller(dev, opts.detach_kernel_driver)));
+      break;
+
+    case GAMEPAD_XEOX:
+      lst.push_back(ControllerPtr(new XeoxController(dev, opts.detach_kernel_driver)));
       break;
 
     case GAMEPAD_LOGITECH_F310: