+++ src/xboxmsg.cpp
@@ -60,6 +60,9 @@ std::string gamepadtype_to_string(GamepadType const& type)
     case GAMEPAD_SAITEK_P3600:
       return "saitek-p3600";
 
+    case GAMEPAD_XEOX:
+      return "xeox";
+
     case GAMEPAD_LOGITECH_F310:
       return "logitech-f310";
 
@@ -101,6 +104,7 @@ std::string gamepadtype_to_macro_string(GamepadType const& type)
     case GAMEPAD_FIRESTORM_VSB: return "GAMEPAD_FIRESTORM_VSB";
     case GAMEPAD_SAITEK_P2500: return "GAMEPAD_SAITEK_P2500";
     case GAMEPAD_SAITEK_P3600: return "GAMEPAD_SAITEK_P3600";
+    case GAMEPAD_XEOX: return "GAMEPAD_XEOX";
     case GAMEPAD_PLAYSTATION3_USB: return "GAMEPAD_PLAYSTATION3_USB";
     case GAMEPAD_PLAYSTATION3_BLUETOOTH: return "GAMEPAD_PLAYSTATION3_BLUETOOTH";
     case GAMEPAD_LOGITECH_F310: return "GAMEPAD_LOGITECH_F310";
@@ -149,6 +153,9 @@ std::ostream& operator<<(std::ostream& out, GamepadType const& type)
 
     case GAMEPAD_SAITEK_P3600:
       return out << "Saitek P3600 (Cyborg Rumble)";
+
+    case GAMEPAD_XEOX:
+      return out << "Speedlink Xeox USB Gamepad";
 
     case GAMEPAD_PLAYSTATION3_USB:
       return out << "Playstation 3 USB";